refactor: move common components to ui/; add industry FreightRoute with CBMCalculator and QuoteForm; introduce UI components (CtaButton, ResponsiveImage, Hero, ValuePropositions) and update exports
